menu code runs in main process
- move menu templating to main-only module
- menu template change can be requested from renderer windows via ipc
- temporary `createResponders` wrapper for all menu features
- menu updates when language is changed in preferences
